Move actors to one module
