--- src/actor.cpp
@@ -1,69 +1,0 @@
-/* File      : actor.cpp
- * Program   : mrtp
- * Copyright : Mikolaj Feliks  <mikolaj.feliks@gmail.com>
- * License   : LGPL v3  (http://www.gnu.org/licenses/gpl-3.0.en.html)
- */
-#include <Eigen/Geometry>
-#include <cmath>
-
-#include "actor.hpp"
-
-
-namespace mrtp {
-
-Actor::Actor() {}
-
-Actor::~Actor() {}
-
-bool Actor::has_shadow() { return has_shadow_; }
-
-float Actor::get_reflect() { return reflect_; }
-
-float Actor::solve_quadratic(float a, float b, float c, float mint,
-                             float maxt) {
-    float t = -1.0f;
-    float delta = b * b - 4.0f * a * c;
-
-    if (delta >= 0.0f) {
-        if (delta > 0.0f) {
-            float sqdelta = sqrt(delta);
-            float tmp = 0.5f / a;
-            float ta = (-b - sqdelta) * tmp;
-            float tb = (-b + sqdelta) * tmp;
-            t = (ta < tb) ? ta : tb;
-        } else {
-            t = -b / (2.0f * a);
-        }
-
-        if ((t < mint) || (t > maxt)) {
-            t = -1.0f;
-        }
-    }
-    return t;
-}
-
-Eigen::Vector3f Actor::generate_unit_vector(Eigen::Vector3f *vector) {
-    float tx = (*vector)[0];
-    float ty = (*vector)[1];
-    float tz = (*vector)[2];
-
-    float x = (tx < 0.0f) ? -tx : tx;
-    float y = (ty < 0.0f) ? -ty : ty;
-    float z = (tz < 0.0f) ? -tz : tz;
-
-    Eigen::Vector3f unit;
-    unit << 0.0f, 0.0f, 1.0f;
-
-    if (x < y) {
-        if (x < z) {
-            unit << 1.0f, 0.0f, 0.0f;
-        }
-    } else { // if ( x >= y)
-        if (y < z) {
-            unit << 0.0f, 1.0f, 0.0f;
-        }
-    }
-    return unit;
-}
-
-} //namespace mrtp
